requirements.txt
Flask
gunicorn
numpy
sklearn
pandas
evidently
dill
PyYAML
gdown
mysql-connector-python
-e .